refactor: organise components for move to analytics
