+++ src/analyticsComponents/Toolbar/HoverMenuBar/index.js
@@ -2,10 +2,9 @@ import { Dropdown } from './Dropdown.js'
 import { HoverMenuBar } from './HoverMenuBar.js'
 import { Menu } from './Menu.js'
 import { MenuItem } from './MenuItem.js'
-import { useHoverMenubarDropdown } from './useHoverMenubarDropdown.js'
 
 HoverMenuBar.Dropdown = Dropdown
 HoverMenuBar.Menu = Menu
 HoverMenuBar.MenuItem = MenuItem
 
-export { HoverMenuBar, useHoverMenubarDropdown }
+export { HoverMenuBar }
