+++ src/analyticsComponents/Toolbar/InterpretationsAndDetailsToggler.js
