+++ src/analyticsComponents/index.js
@@ -1,4 +1,1 @@
-export * from './HoverMenuBar/index.js'
-export { InterpretationsAndDetailsToggler } from './InterpretationsAndDetailsToggler.js'
 export * from './Toolbar/index.js'
-export { UpdateButton } from './UpdateButton.js'
